requirements.txt
lightning
hydra-core
transformers
sentencepiece
wandb
evaluate
rouge-score
bert_score